business_acquisition_system-main/agent_3/requirements.txt
langgraph>=0.0.20
langchain>=0.1.0
anthropic>=0.18.0
pandas>=2.0.0
python-dotenv>=1.0.0